Fix classloader problem which occurs when a war file gets expanded for a context that is already listed in the server.xml.
The classloader gets created by server.xml processing before the war file
is expanded.  When SimpleClassLoader is used, it will lack the WEB-INF
directories in its urls[] array.  Make AutoSetup remove and add the context
so the classloader gets updated.

Fixed the indentation to make it more readable, though it somewhat hides
the actual changes.