Fixes and changes for JDK1.1 support.
The build on 1.1 was broken for few weeks, and I hope this solution
will improve the things a bit (in addition, I'll check the nightly 1.1
build until we're done ).

The util.compat package contains a helper that will avoid the use of
introspection or other (complex) solutions to maintain 1.1 compatibility.

It's just a small wrapper around common methods, with 2 implementations -
one for jdk1.1 ( doing nothing for doPriviledged, returning SimpleClassLoader,
etc) and one for jdk1.2+ ( wrapping and performing PriviledgedAction,
returning the URLClassLoader, etc).

As a result:
- LoaderInterceptor12 is no longer needed
- LoaderInterceptor11 will automatically use URLClassLoader if in Java2
environment, no need to edit server.xml for that
- Jdk12Interceptor compiles and can be loaded in jdk1.1. It'll do
java2 tasks if java2 is detected ( there is another trick to add it only
if needed - see next )
- and of course, it builds and work fine on JDK1.1 ( again )

As a related change, PolicyInterceptor will be automatically
loaded if Java2 is detected and a security manager is set (or a
secure option is set on the contextmanager ). To do that we use
PolicyLoader, that detect the conditions where we want sandboxing
and if java2 is available and loads the "real" policy interceptor.

That simplifies a lot the configuration, we no longer need to explain
how to change the config just to run JNDI in java2 environment.

I also added the permissions needed to run /admin in a sandbox, there
are few tests that are failing in sandbox - I'm working on a fix
( but probably next week, I have a deadline for monday )

Also, added a small comment ( to be expanded ) in BaseInterceptor.