Updated GTestTag to process "port", "http.protocol", "host" parameters and set them as ant properties. This allows testing with apache ( where http.protocol is HTTP/1.1, and the port is different ), and allows to test virtual hosts.
Added the options in the simple form in test.jsp. ( again, the graphical
aspect of the test application sucks, and most of the taglibraries used
are ugly - the goal is to find and fix the bugs, we can make it look
nice later )

It would be really great if someone can "beautify" the test page and the
other admin pages, etc - but it's low priority right now.