- removed unused attribute from <ant:ant>
- fixed the auth-constraing ( the role in admin-users.xml is tomcat_admin )

- update watchdog-servlet.jsp and watchdog-jsp.jsp to use <ant:ant>

( all 3 tests are working fine now - and passing. I found the problem with
watchdog jsp failures - it seems that one of the watchdog taglibs is
using a static variable that is initialized at compile time - so if
the .jsp is not re-compiled the test will fail. The only workaround is
to do a "rm -rf work/DEFAULT/jsp-test/jsp/tagext )