The first significant change in core for this year :-)
The startup rules for tomcat were almost undefined in tomcat 3.2 and before,
and defining a clear state was one of the biggest issues for 3.3.
Things has improved ( thanks to Larry and Nacho for all the feedback),
but I don't think they are perfect yet

This is one of the biggest things that need review and attention.

This fix tightens up a bit more the startup process:

- no addContext or initContext hook can be called before the server is
fully configure ( i.e. in STATE_CONFIG ). Because the server.xml reader
adds both modules and contexts, calling addContext is delayed until
init()

- make clear that after init(), the server is fully configured and initialized
engineStart is mostly for connectors.

- reorganized a bit the documentation - runtime configuration is a different
issue.

Please read at least the ContextManager comments on the top, and send
any feedback.


( ContextManager has 1 very important role - call the hooks that
compose tomcat. It also stores global config, modules and contexts -
but that's easy )