Started work on 427 - if a war file is specified in server.xml in a different location than webapps, we should expand it.
( I'm not sure this is the right thing - IMHO the right thing would be to
do a simple deploytool - even as a simple shell script or in ant - and
do the deployment the right way. Auto-expanding war files was just a
simple solution to the lack of a real deploy tool. )