Few changes in Context to make sure the state and rules defined in ContextManager are respected. ( i.e. no add/initContext before all modules are initialized, etc ).
Note that per-Context modules have less power than global modules,
i.e. engineInit is called only when the Context is added, contextMap
is never called, etc.

That means that "configuration" modules and "context mappers" need to
be "global". It would be a good idea to make all modules that can be
context-local part of a "profile" to simplify the configuration
( for example a set of contexts that may share common authentication
style, logging, etc - right now you have to set it on each module )
Global modules are bad because it's hard to tune them for individual
contexts.