All parameter handling code is now part of Parameters. The internal representation of parameters is now encapsulated, no external class can access the internal details.
This is still the old code and representation ( hashatble of
name-> value[] ), but a the processing has been improved and now follows the
rules defined in the spec for merging GET/POST params and merging
RequestDispatcher queries.

It'll also simplify RequestDispatcher a lot by removing all the hacky
code to support parameter saving/restoring.

( this started as a fix for JspInterceptor to deal with jsp_precompile,
but it's better to do a real fix instead of a simple hack. )