Fixed the JDK1.1 build problem.
It's very interesting - the code compiles fine on JDK1.2 because getParent()
is now part of ClassLoader.

It doesn't in JDK1.1, but adding a getParent in SimpleClassLoader will work
only in JDK1.1 - in 1.2 it'll fail ( it's final in ClassLoader ). So we have to
call the method getParentLoader() in SimpleClassLoader ( so it compiles in
both ) and abstract if via JdkCompat.

But we are lucky to have it behind JdkCompat and not have to worry about this
in future.

( I also followed the XXX and moved SimpleClassLoader in compat - it shouldn't
be used standalone - you want to use the real thing if available -
URLClassLoader )