Fix for the destroy() bug reported by JDNovotny@lbl.gov.
Now both jspDestroy() and destroy() should work ( please check the next
nigthly build ).

The fix is specific to 3.3.

( BTW, I can try to do a fix in 3.2.x, but if the bug is present in 3.2.1
and it's not a regression I would rather fix other bugs in 3.3 )

The problem was that the servlet was removed from context, and then
destroy() did nothing ( it checks if the servlet is initialized - when
it's removed from the context the servlet is no longer in a "stable"
state, it's orphan )