Fix the <html:form action="xxx"> attribute to correctly work for both extension mapping and path mapping of the controller servlet -- at least as long as your servlet container correctly allows the following to work:
	InputStream is =
	  getServletContext().getResourceAsStream("/WEB-INF/web.xml");

like it is supposed to.

The new recommended approach is to set the action attribute of <html:form>
to exactly the path of the Action to which you wish the form to be
submitted.  The <html:form> tag will understand how the action servlet is
being mapped, and will construct an appropriate URL accordingly.

For backwards compatibility, the existence of the extension-mapping
extension on an action attribute still works, so either of the following
tags would work in the Struts example application:

	<html:form action="/logon.do" ...>

	<html:form action="/logon" ...>

The latter path will also work if the controller servlet is path-mapped,
but you will need to adjust the paths for several of the <forward>
elements in this case.

Submitted by:  Marcus Ahnve <marcus.ahnve@lecando.com>
PR: Bugzilla #587