util/log changes.
As proposed, the actual implementation is now in util/qlog, and
util/log contains only the base classes.

Also, the code was slightly re-organized - the log registery is now
part of LogManager ( which is of interest for the LogModule that
sets up the loggers ), LogHandler ( which is of interest for the actual
logger implementations - for example if you want to write an adapter for
log4j ) and Log ( which is the only class used in tomcat, utils, modules ).

The 3 classes in util/log are functional ( i.e. if you don't set any LogHandler implementation you just get the messages on System.out ).

The normal use is:
Log.getLog()
log.log(...).

Normal code shouldn't deal with anything but Log ( the other 2 are for
logger implementations and for configuring the logger ).

The code is functionally identical with what we had previously, and
AFAIK doesn't have anything but the core functionality we need.

BTW, I added few comments about the reasons to have a 2 level logger and notes
on security ( i.e. normal code shouldn't be able to change log config )