Port update to SessionExample servlet from tomcat_32 by Marc Saegesser
We now include some additional information about session handling, such
as whether the session is new, what the request session id was, if the
requested session id was valid and the source of the requested session
id.  There is also a new button for forcing the current session to be
invalidated.

These changes better demonstrate the session management features of
the servlet API and also provide a way to test the recent patches
for session id handling.