Port changes from tomcat_32 by Marc Saegesser
This fixes some additional problems uncovered by the fix for
Bugzilla 160.

The isEncodable() method used isRequestedSessionIdValid() to determine
if there was an active session.This is incorrect, because the requested
session id may have expired or been invalidated and a new session created.

isEncodeable() now encodes sessions that are new (i.e. we don't know yet
if the client will be sending cookies or not) or if, if the session
is not new (meaning the requested session id was a valid session) and
the requested session id did not come from a cookie.