My last change for M2 - replaced the attempted get/setNote with get/setInfo.
The hook is intended for lazy evaluation of request informations like
the encoding and for more flexibility in attributes like the SSL chain.

Tomcat performance is due in part to the lazy evaluation - certain fields
in request are not used in all requests, and pre-computing them is a waste.

There is another problem - supporting attributes like the SSL chain
no longer requires the SSL module to set an attribute ( i.e. the chain ),
but can be delayed until and if the request needs it.

In the case of encoding - while this doesn't change too much in the
current implementation, it allows modules to provide smarter mechanisms
to detect the charset and more optimizations.