First attempt to fix the UTF decoding bug.
This doesn't change the behavior - what worked before should work the
same ( i.e. bytes<0x80 ). For chars > 0x80 ( that didn't worked with
the previous code ) I cut&pasted code from xerces.

The idea is to avoid using the (memory expensive) String( bytes[] ),
which proved to be a performance problem in many cases.

Xerces is doing exactly the same thing to optimize the conversion.

This works for UTF8 - for all other encodings the slow method is used.
There is another way to resolve the problem ( a trick similar with
OutputBuffer ), and we may implement it someday, but the hope is that
people will start using unicode as a clean and simple solution for
non-ascii charsets.

Netscape and IIS and most platforms already support UTF. ( that doesn't mean
we shouldn't deal with the other encodings )

( the code is untested - I need UTF examples, adding files in
various encoding with "strange" names to the test suite is still on the
todo list )