Test for Bugzilla#291
Using an html file as the <location> in an <error-page>
in web.xml causes a StackOverflowError. BugRat Report#538

Reported by: Catherine.Jung@netcentricsystems.net (Catherine Jung)

Actually the fail obtained in Tomcat 3.3 is not the one reported,
this report was for 3.2.1

But it's actually a bug in the error handling in TC3.3, when a error is redirected
to a html page, the static interceptor actually gives a not found error (404), because
the passed request is the old request uri  ( the one of the failing servlet )
not the html  error page,  thus giving a 402.

i will try to reproduce  it on 3.2.2 too.