Port tomcat_32 security fix by Craig McClanahan
For Tomcat 3.2, fix the security vulnerability reported by Hiromitsu
Takagi.As with Tomcat 4.0, the problem is not related to JSP
specifically.It was caused by the fact that the original request URI was
included in the standard error page produced by Tomcat for errors like 404
(not found), which was the illustration in this case.

WARNING:Web apps that echo the request URI in their responses (either in
a standard response or in an error page) can be subject to this same kind
of vulnerability.