made changes to TR.props to get service initialization in correct order and added props related to pull tools.
rearranged build.xml a bit, so that turbine and torque are checked and possibly
built first.   Torque is now building into the src directory and then copying
the generated files to the build directory before deleting them.  This makes it
easier to add tables.

updated some jars/zips to the latest versions.

working on the Report.vm template to populate selects and made minor changes to om classes while working on the template.

PR:
Obtained from:
Submitted by:
Reviewed by: