I'm looking into session expiration bugs and I decided it's easier to split the expire code so we can read it and understand what happens.
I also created a separate module for ID generation, as the code is
independent of the session storage ( which we want to keep it simple
in case someone wants to backport file store or other advanced
session stores ).

In the expirer, the main difference is that we use a single expire thread
per context manager - instead of one thread per context. This allows
tomcat to scale a bit better when many contexts are deployed
( some dynamic loading/unloading is needed, but that can be done later ).
( there is no gain in having 100 expire threads working in paralel -
even if you have a 100 processor machine, you don't want expirer to
take all the time. You can still have multiple expirers if you explicitely
define it per context )

I tested the code, and it seems the sessions are expiring fine. I'm
still working on the admin pages that will show the session statistics
and allow a bit of control.

Larry - please don't kill me, the changes are not very big ( even if it
looks like ), I don't think they add too much instability but makes
the code clearer.