PATCH: the security role references were never mapped to security roles in the EnterpriseContextImp.isCallerInRole() method.
For now, all references are passed on to the RealmMapping implementation
regardless of whether the <security-role-ref> element is found in the
deployment descriptor or not. If the security role reference is not
found, a warning is logged by the server. A more strict compliance may
be enforced in future versions.