Temporary(?) fix to stop crashes when creating a new Session in embedded use of Jasper (i.e. as standalone JspServlet ).  Problem may occur in other situations as well, I'm not sure.
Problem:  When SimpleSessionStore.SimpleSessionManager.getNewSession() is called (as a result of a call to request.getSession(), the session ID of the new session has not been set.  Since it is implemented as a MessageBytes object, the .toString() method to get a string representation of the session id returns null.  When getNewSession() tries to use this value as a key into the sessions hashtable, a NullPointerException is thrown.

Fix:  To stop the crashing, I've added a line to "getNewSession()" to replace a null value for the string representation with the literal "null".  This should NOTbe considered a fix of the underlying problem:  Why is newId == null (i.e. why is the new session's id not initialized)?  The implication is that request.getSession() is broken in jasper right now, at least when using Jasper independent of tomcat.