improved build.xml so that torque does not regenerate for all turbine changes.
Added a property to identify a scarab instance.

added 3 columns to SCARAB_ISSUE, which can serve as a unique_id.  The intance_id
comes from the property file and is left null in the case the issue is local, the id_prefix comes from the project the issue is entered against, and finally a
counter which is generated through IDBroker.

Added a MODULE_CODE to modules that can serve as the prefix for issue ids.
(Not implemented - but will need to be:
When adding a top level project a unique module code must be entered.  Component
level modules can default to their parent code or have their own.)

removed some repetitive code from Issue since a method was added to Module.
Implemented the id generation when the issue is first saved.

removed some irrelevant html from Wizard2.vm, but functionality is not there yet.


PR:
Obtained from:
Submitted by:
Reviewed by: