Patch from  Bojan Smojver <bojan@binarix.com>, slightly modified.
The patch allows systems that have /dev/random to use it instead of the
slower Random. Instead of checking for OS==linux ( as in submited patch )
we use an option of the module.

The code if the option "useDevRandom" is not set is the same as before.
If you set useDevRandom="true" then /dev/random will be used.

( Bojan - please review and let me know if it is not what you intended,
I'm not sure if the /dev/ransom needs to be closed/open all the time )

Submitted by:	 Bojan Smojver <bojan@binarix.com>