ok,
the values that are already there can't be moved to a number that doesn't
exist.

ie:

assume 1 based (instead of 0 based):

if you have 8 attributes, then they should only have popups for 1-8
and the new popup should be starting at 9.

this patch fixes the above.

the only other bug that i see right now is on the reordering. say you set
8 to be 5 and you click save, it comes up as 6 instead of 5. this seems to
be an off by one bug in the Action somewhere...not really life threatening
though.

-jon