added a couple join tables to allow modules to specify the options and to specify a hierarchy in the options.
The module_option relationships are pretty much working, so that it is still
possible to enter issues.  The attribute editor may be broken and it
set up to edit attributes on a global basis.  A module_option editor
will need to be built.

The option hierarchy is still not functional.


PR:
Obtained from:
Submitted by:
Reviewed by: