Check in for 1385, from bojan@binarix.com (Bojan Smojver).
Any random file can be specified. The default ( in server.xml ) is to use
/dev/urandom if it exists. To minimize user configuration we do set it
in server.xml and let the module decide if it exists or not.

The difference in speed is quite significant - instead of waiting 4-5 seconds
on the first JSP page ( or servlet with session ) now we have instant response.

Submitted by: bojan@binarix.com (Bojan Smojver)