Adding the (slightly refactored) runtime of jasper ( the tomcat3.3 version ).
Changes compared with the original:

- new package names

- all runtime in the same place ( Constants and JasperException moved in
runtime, since they are runtime-used )

- Constants no longer have "log" or "getString" methods, but only constants

- no dependency on any external code !!! No logger, simple pool or string
manager - those services should be provided by the container ( or container
adapter ), and a Liaiason interface is provided.

- JspFactoryImpl is the class used to plug in the Liaiason ( as sugested
in the specs ). From an jasper integrator view it's the only object it
should manipulate ( next step: add explicit setters for all things that
the runtime know, stop using "tricks" like runtime properties. The integrator
should provide configuration of JspFactoryImpl via whatever it want.

Note that we still have a lot of code that is not actually used at runtime
but at compile time - that should go away in time.

As soon as the compiler is modified to generate self-contained code we should
be able to include the runtime in the war, with jspc-generated code and
have self-contained apps ( from the point of view of jasper ).