Modify RequestUtils.absoluteURL() to throw a MalformedURLException if one is encountered while trying to reconstruct a URL from the specified page context and context-relative path.  This will cause tags like <html:link>, <bean:include>, and <logic:redirect> to throw an exception instead of create null URLs.  In addition, these three tags will log the exception to the servlet context's log to facilitate further analysis.
These changes are in direct response to Bugzilla #1586, where the main page
of the example application creates bogus hyperlinks that look like this:
  <a>Log on to the MailReader Demonstration Application</a>

I suspect that this is due to misconfiguration of the servlet container
being used (WebSphere in the case of the bug report).  Throwing the exception,
and logging the details, will prove it one way or the other.