moved the pick a module hack to ScarabPage so it does not have to be added to every Screen and Action method, this will be easier to remove later as well.
having the ScarabRequestTool return ScarabUser objects, so that
we do not have to always cast it.

added the ability to set parent-child relationships in the
options editor.  Probably not the best interface, but easy to
implement and will likely be blank for most attributes anyway.

worked more on the ModuleAttributeEditor, partially functional, at
least what I tested worked.

PR:
Obtained from:
Submitted by:
Reviewed by: