Change RequestUtils.computeURL() back to returning a String -- the java.net.URL class does not like dealing with page-relative or server-relative URLs.
When a "forward" or "page" attribute is used to specify the hyperlink, a
server-relative URL (beginning with the context path) is generated
*without* the scheme, server, and port portions.  This assists in many
cases where proxy servers intervene between the client and the actual
server.  (If an "href" attribute is used, it is passed through unchanged.)

I haven't addressed the problem with computeParameters() yet -- that will
be next.

PR: Bugzilla #1356
Submitted by:	Stoehr Sukachevin