Added a generic Byte->Char (charset) decoder, using the same model as the C2BConverter used by OutputBuffer.
The input system is still very inefficient, this is just the first step
( not used in active code - I did a lot of tests but it's better to be
safe for now, and make sure we don't have regressions with the other changes
before activating this one ).

UTF8Decoder used to be part of ByteChunk, it was fixed and tested with
multi-byte chars.

The decoding was refactored for 3 reasons:
1. Performance. This way we can optimize individual converters and plug
specialized versions ( like UTF8 )
2. We should be able to use ( when detected ) nio converters
3. Clarity. It's ( IMHO ) easier to follow the Chunk code without all the
extra ( and complex ) decoding code in.

Both clases are not in use in the current code, but should be enabled
for 3.3 - the code is stable and efficient.