Merged the resizable and flush-able buffer code from OutputBuffer.
Now the CharChunk and ByteChunk can be used as general-purpose buffers
( instead of StringBuffer for example ) - i.e. both input and output.

Most of the code is old and well tested ( has been part of OutputBuffer for
a while ), we just moved it here to be of more use.

We preserved all the old API, except for url decoding ( which has been
almost completely rewritten ).

Some methods are now deprecated and could be removed ( we haven't frozen the
API - if we ship 3.3 with those in it'll be hard to remove later ).
The caching of header conversion result should be part of MimeHeaders, and
not affect the buffers.