Added 2 very common charset detection schemes.
The most important is the per session charset - all browsers will use the
same encoding for the posted data.

Another common one is using an attribute in the request uri ( or parameter -
this is not done yet ).

Both will be disabled in the default config, but I think will be very
usefull for anyone running sites with non 8859-1 charsets.

More to come, please provide feedback.