Fix for JspServlet support.
Tomcat must return a classpath that can be used to compile JSPs ( if
JspInterceptor is not used ).

The problem is that we want to allow the loader to be customized - the loader
interceptor has the class path ( context.getClassPath() returns the local
classpath, not including parent loader's path ).

The soltuion is to use the provided hooks to let modules provide the
required information. This will also alow more customization in the attribute
name, and in enabling/disabling the feature ( if JspInterceptor is used there
is no need to expose this non-standard information ).

This also cleans up "getSpecialAttribute" ( we might want to move the other
special attribute - facade - to a special module ).