changed canPassivate signature back to original canPassivate(ctx).
Sorry, it was a bad decision to begin with...

In passivation, do check for CacheKeys not being equal within AbstractInstanceCache
instead.  In the previous version, if CacheKeys were not equal, PassivatorJob
would check to see if canPassivate was false, if so, re-insert the ctx back
into the cache.  This was a no-no.