do not have to use minLength=1 as a substitute for required any longer.
the setRequired in Login is not needed, as it is meant to be used in cases
where it is only known at runtime if the field is required.

also need to authenticate the user in a try block in the event an invalid username/password combo is entered.

PR:
Obtained from:
Submitted by:
Reviewed by: