Issue entry works.  The Wizard2.vm (2.1.1) still needs some work in that the column sort is not done and voting is not implemented awaiting comments.  I do not know how useful the sorting capability is in 2.1.1, but it will be useful in larger search results, so it will be implemented next.
modified the properties for file upload to match column size and provide some efficiency.

Added methods to ReportIssue to handle some of the extra buttons in the latest ia/ui.
Also modified so that the parameters are not getting passed along during the
wizard, but are saved with the issue that is being entered, The objects are
already consuming memory, the values should not be much more.

It appears some Attributes and Values are going to get hardcoded into the ui,
so I added some constants to handle the ones that have appeared so far.


PR:
Obtained from:
Submitted by:
Reviewed by: