fixed up IssueView attribute editor.
need to have $intake.declareGroups() towards the end of form for intake to work properly.
Also using the SequenceHashtable correctly to maintain the order.
(Issue.getAllAttributeValues will need to be moved to use SequencedHashtable
to maintain order as well.)

Issue.save() was dumping changes in attributes during the check for null values.Should just get the AttributeValues already present in the Issue, do not reload from db.

the tempWorkaround method call occurs in ScarabPage, so it does not have to be in every action any longer


PR:
Obtained from:
Submitted by:
Reviewed by: