Decouple the Hook ( and the core ) from introspection, open it for better ( declarative or code generation ) mechanisms.
The current implementation will still use the introspection ( until
modules.xml will include the correct declarations about the callbacks ).

As a side note, I repeated an old benchamark to find out if "too many hooks
will affect performance"
Surprisingly for me ( I didn't expected that ) - it does, with about 5%
If you disable the code that sets up the hook chains with only the
modules that are actually involved ( that's avg 3-4 calls per chain, most
have 1-2 ), and if you run ~100.000 requests ( few times ), you may find
 a statistical difference ( now that we cut most of the hotspots from
tomcat - a year ago it was none ).

The surprise is that I expected the JIT to just eliminate the empty calls -
5% is not insignificant.
( of course, this is just theoretical - there is no reason to
add all the modules in all chains - that have been resolved a long ago, but
it's interesting to know that it had such a small impact. )