Some more improvements to ApacheConfig, many thanks to William Barker. This is based on his patch sent on tomcat-dev, with few changes.
There is still a lot to add before we could turn this on ( and probably
the most important short term thing is to detect the apps where this
kind of configuration can be used, and use "forward all" for apps where
settings in web.xml do not map to apache configs).

I hope William can help with more patches like this one :-)

Also, now we have an option to generate either JkMount ( i.e. jk_uri_map
will do the work and mapping ) or <Location> and SetHandler directives,
and let apache mapper do the mapping ( probably better ).
This later option doesn't work for the old mapper, and hasn't been backported
to mod_jk for Apache1.3 ( I tested and commited the code for 2.0, when
it's done I can do the backporting but it's too painfull to work with
2 versions at once ).

For IIS the situation is really bad, there is a bit of hope for some
webapps in NES - but for Apache I think we can cover almost all web.xml
settings using the native apache configuration.

Ajp14 will improve a lot the cases we can't cover - both autoconfiguration
( which seems to work just great ! ), and also by protocol extensions that
are planned for jasper ( and will help a lot in static content handling ).


Submitted by:	william.barker@wilshire.com