This should fix some of the classcastexceptions i have been getting as a result of bad references to the wrong objects: 	made ScarabUser an interface 	renamed ScarabUser -> ScarabUserImpl 	updated the schema.xml to reflect above changes 	ScarabUserImplPeer replaces ScarabUserPeer
Modified a bunch of code to reflect these new changes and to use the
interfaces instead of the objects. This includes finding and fixing lots
of references to Module -> ModuleEntity...there are still more out there,
but this is another good sweep of the code.

One thing I did that I don't like is that
Issue.getModuleCast(ModuleEntity) needed to
be created in order to cast to BaseIssue.getModule(Module)...this is
unclean, but i'm not sure how to fix it quite yet because in the Base
objects there are still tons of references to Module instead of
ModuleEntity. We need to make Torque (or the schema.xml) more intelligent
so that it generates the right references to the right objects.

Happy 4th.

-jon