Make sure modules.xml file is configurable.
Another small optimization in startup time, by caching the information
we need in modules.xml and avoiding a parse.

( modules.xml is supposed to grow in future to allow much more decriptive
info about tomcat modules, most of it will be used by the admin interface,
tomcat only need name/class name - sort of taskdef in ant )

This can be turned off as default ( but I don't see any reasons for that ).

A next step would be to cache the result of Hook introspection, and maybe
( a bit more difficult ) the XmlMapper work can be avoided - that should
reduce the startup time to an insignifiant value.
( and of course, allow very small tomcat runtime, without xml or introspection)

I'm not planning any of those for 3.3, but in a new set of config
modules ( that could be used instead of the default one ), after 3.3 is
released.