Migrate Struts HEAD branch to use the Commons packages (beanutils, collections, and digester) rather than the Struts internal versions of these classes, as discussed on the JAKARTA-COMMONS mailing list.  In nearly all cases, this was as simple as changing the import statements.
WARNING:  During this conversion, an anomaly was discovered with the
Commons version of Digester (which supports JAXP/1.1, and therefore can be
namespace-aware).  At the moment, the Commons version does not match
nested elements correctly *unless* you are running in namespace-aware
mode.  This is a bug in the Commons version of Digester, which is
currently worked around in Struts by calling

  digester.setNamespaceAware(true);

on any newly constructed digester instance.  Applications who create their
own digesters, and are migrating to the new classes, will need to do the
same thing.

WARNING:  The Commons version of Digester depends on JAXP/1.1 rather than
JAXP/1.0, so I will also be updating the prerequisite software
requirements.  Two parsers known to work successfully are the JAXP/1.1
reference implementation from Sun, and Xerces 1.3.1 from Apache.