Ops... Normalization _after_ decoding, of course - thanks Larry, I can't believe I didn't realized that...
One problem that remain is with the "unparsed" URI. Internally we deal
only with decoded and normalized URIs, that includes the JSP interceptor and
all tomcat modules.

Servlets, on the other side, get the original request ( as required by spec).
We could normalize it ( not allowed in the spec, but better safe then
compliant :-). That wouldn't solve anything - if we normalize undecoded
url we can still have the "." and "/", and if we decode - then we completely
brake the spec.

On the other side, if the servlet needs to do anything with the URI, it has
to decode it and implement all the security checkings itself. Not easy, but
it's not our problem.