Remove the SetHandler option.
This is new code, and will probably be available with all the new
features and ajp14 as part of j-t-c.

There is no need to put it into the official 3.3 beta/release - people
who need this feature will be better served by a standalone module, as
this is likely to evolve in the future ( as oposed to the 3.3 "main" codebase,
which is supposed to remain freezed - so we can focus on features and modules
knowing the base is stable and it's not going to change )

The current autoconfiguration mechanism in 3.3 ( using JkMount ) should work
as well ( or better ) than before, with either the old mod_jk or the new
one.