Fix for the fix.
The original fix allowed modules to be defined in server.xml ( or equivalent ).
If the modules are loaded from modules.xml, the loader will automatically
add the new rules.
If the cached properties files are used - we need to add the rules explicitely.

The "modules" note is also used by ContextXmlReader, to avoid parsing the
same file multiple times.