Few fixes and usability enhancements to JNIConnectionHandler.
The native library is loaded only if we're in "native" mode, the module
is inactive otherwise. This allows us to keep it in uncommented in server.xml.
( just like Jdk12Interceptor - which detects if jdk1.2 is used )

Also, the module needs no parameter if the default location for the library
is used ( i.e. TOMCAT_HOME/bin/native/jni_connector.[so,dll,nlm] ). The
extension is computed based on os.

A message is displayed if the file can't be found.