Add a variation of the existing isTokenValid() method that optinally resets the token in the same, indivisible, operation.  This avoids a potential race condition that bypasses the check made by the transaction token, as follows: * User POSTs a form * Action finds the form valid and starts processing, without resetting   the token at this point * User presses STOP and submits again * Action (on a second request thread) also finds the token valid   (because the first thread didn't reset it).
Note that, even if you call resetThread() immediately after
isTokenValid(), there is still a (very small but finite) chance of a race
condition with the old version.  To avoid that, you can do a "test and
reset" atomically by calling:

  boolean ok = isTokenValid(request, true);

Note that if you do this, you'll only be able to call isTokenValid() once,
so save the result if you care about it.

For backwards compatibility, calling isTokenValid(request) behaves as it
did before, which is (in effect) the same as:

  isTokenValid(request, false);

Many thanks to Yuhong Guo for explaining how this could happen, and
suggesting a solution!

Submitted by:	Yuhong Guo <YGuo@infinet-inc.com