Small changes ( but the diff is big ). I tried to put in order all the package/protected/final stuff.
Having final methods and classes hava a tiny impact on performance but may
affect our ability to do "hot fixes" via add-on modules, and to extend
tomcat or tune it for special configurations.

Let me know if this creates a problem for anyone - I believe removing the final
doesn't affect in any way the code stability but provides a lot of flexibility.

I also added few factory methods in CM, to completely allow specialized
extensions to control the whole thing.