A much more serious problem here. We recently fixed DependClassLoader - it needs to use defineClass() itself, it it relys on the parent the other classes which depend on the loaded class will be loaded with the parent loader.
This change had a side effect - since defineClass() was used with 3 parms, without
the protection domain.

The first fix was required to fix reloading, this one is required to fix sandboxing.
I reproduced the same tricks we used in jasper to maintain compatibility with JDK1.1
( however, I may need to add one more method ). This is not finalized, and is
possible it'll brake JDK1.1 compilation ( not difficult to fix, but I have few other
changes to commit).