Ok, this is the big one !
I did a lot of work on this area, now things should be much cleaner and simpler
( by my taste ).

First, most of the settings from Main.java are now part of EmbededTomcat. We
expect people to use EmbededTomcat to include tomcat in other apps, not
Main - since we need to provide a reach interface and more control.

Main.java is now reduced to the same role as "java -jar" ( and is equivalent
with it - or should be ). It just create a simple class loader including all
the classes in common, and then pass control to EmbededTomcat, which does
all the real work.

In other words, to embed tomcat you need to include all the common in
a classpath, then use EmbeddedTomcat as a regular bean - set properties and
call the methods you need.

EmbededTomcat will set the container environment ( which must be separated
in another class loader - security and isolation ).

I also added a bit more documentation.

An important change - there is no dependency between ET and container classes
( see 2 lines above ).

EmbededTomcat now  works in both modes - if setEstart is used ( equivalent
with java -jar tomcat.jar estart ), it'll use the hardcoded modules
( or an alternate set, depending on your code ). The default, or if
"start" command is used on the command line, is to load tomcat using
server.xml, as before.