code cleanup in Attribute.java...john, please proof changes for MT issues. there is still more work to do here...
fixed scarablink from adding template to the path multiple times.

more work on the AttributeEdit templates...didn't get very far cause most
of the day was spent going over scarab code and getting a feel for the way
that things are being done...

note: we need to fix the issue with regards to tracking a module via a
user...this is a bad assumption, it should be tracked in the URI...today's
checkins should make this more easily possible. also need to add code to
the templates to validate that the module is a valid module.