man when will this ever end?
did I tell you about the time I had a dream of great fields and sun in the field.  I was stoned I think, but it is irrelevant, the point is that I really enjoy this mix of technology and vision the heady stuff and the grungy.  I really think that "design" is a dead practice for people who need help.  As long as we can stay on the fire we will adn advance this baby of ours. and what else... nothing much...

ah yeah this is really to see who is reading all these commits, the first motherfucker to yell "Kruder and Dorfmeister are pretty cool but they should go easy on the bug powder dust" gets a free JBoss underwear signed by moi.

no one said it would be easy, relax your mind and give it all to me...

you know and I know... our love is strong

to weather the rain
to weather the snow
to weather the snow

Often I see fear in your eyes.
And sometimes, I know your heart is full of lies

trust in me and no one can do you wrong

cause I know and you know our love is strong.

to weather the rain
to weather the snow
to weather the snow

--LAMB, trans fatty acid


alright, :wq! me now, please