changes to query page UI...
couple of important things here. made all the date range form elements pop-up menus... note the "hour" pop-up.

also, personnel search. this is not "hard wired" as we discussed yesterday, but I recognize an enormous oversight in not calling it out in the 2.2.0 query page (and on the saved query page 2.2.1)

Brian