Fix "Usage"display for Win9x.
Remove -security option in favor of using "sandbox".  For start and run,
always set java.security.policy so "sandbox" can be any argument.  This
make it consistent with tomcat.sh.  If the second argument after "start" is
"sandbox" or "-sandbox', still display a message and set a title if NT or W2k.