Added more comments, more logging.
Recycle the facade - this is important because prevents user code from "keeping" references to
session objects that are reused. Since session is a long-lived object the impact on performance
is minimal ( and can be avoided by pooling the facade at context level - if we want to ).

There are 2 more changes needed to support that ( recycle() must be called _after_ the session
is in EXPIRED state - various hooks in expire need access to facade, and we don't want to
depend on module ordering )