Make tomcat.sh consitent with tomcat.bat, and better use of the TOMCAT_INSTALL/TOMCAT_HOME.
TOMCAT_INSTALL is the directory where tomcat is installed, where bin/ and lib/ are located.

TOMCAT_HOME is the local instance, where conf/, webapps/, work/ and log/ are located.

Larry did the cleanup on the .bat file, this is just cut&paste ( the shell script looks a bit
more complex because we're also following symlinks in case tomcat.sh or a link to it
is in path - so the lazy user doesn't have to set the env )

( well, I'm a lazy user as I never set TOMCAT_HOME, too many instances on my machine :-)