Extra check, again result of a report from Bill Barker <wbarker@wilshire.com>.
The name can become null if we have a conversion error ( i.e. the param contains invalid
chars in the detected charset ).  Bill reported a more complex problem, with charset having
a wrong value - but this is a corner case and I doubt there is anything anyone can do
( if the browser doesn't send the charset and the user changes the encoding
during the session and it does so without first sending a page with the new encoding,
I don't know any way we could guess the new encoding )