Number 7 on the todo list, keep DateFormat as a local variable in the response.
The request/response is recycled and there is only one HttpResponse per thread - so
no synchronization is needed.

Initial fix for #1 ( I'm refering to Larry's "Remaining issues" ), it needs to be verified. I
tried to avoid sync() on the typical case, and have the overhead only if the value implements
the listener.