X509Certificate is not available in JDK1.1, moved the code to Jdk11Compat and make sure everything is again working in 1.1.
As ussual, the trick is to have a method that does nothing in Jdk11Compat and
override it with a real method in Jdk12Support ( the names are pretty bad ). This adds
almost no overhead in jdk12 ( just a virtual method call ), and is pretty clean ( as
oposed with other solutions using introspection or other tricks ).

I also added the catch(Exception) in JspReader - I think it's a bug in gcj, but it's an easy
workaround ( gcj is also JDK1.1 AFAIK )