Some fairly major changes:
ModuleEntity.getName now returns a unique name according to the needs of
the Group interface

ModuleEntity.getRealName now returns the specific name for a ME

removed the unused RMouldeUserRole table and .java files because we are
using the TURBINE_USER_GROUP_ROLE table instead

added PickModule which allows someone to choose the module they are
currently working on. thanks to Gary Bartlett <gary_a_bartlett@yahoo.com>
for the initial contribution.

-jon