Another small usability improvement.
Modules ( == trusted contexts, including tomcat-add on services ) can now add classes to
the container and common loader. That eliminate a lot of configuration effort and provides a
very simple mechanism to add libs, using regular war files.

Note that nothing is changed in tomcat operation if no module is included ( the current
case ), it should operate exactly as before.