a whole boat load of changes:
#1. cleaned up use getting the tools in actions by adding helper methods
to the base classes

#2. added an Issue | Search permission and fixed some of the
permission/module checking in the Default.java screen.

#3. cleaned up some areas where ScarabModule was being imported.
John/Elicia please take note of this and study the commits. please don't
import ScarabModule...only the interface (ModuleEntity) should ever be
referenced.

#4. this is the first step towards also adding security to actions with
regards to making sure that scarab.getCurrentModule() doesn't return
null...we aren't there yet, but now it will be easier to use find/replace
to do the work. will probably rename the RequireLoginFirst action as a
result.

#5. added FIXME notes where appropriate...there is a couple areas where a
User object is retrieved directly from the database. this is bad with
regards to integration issues. need to figure out another way to deal with
this through API's. the problem is that the API's currently only have a
way to retrive a user by name yet the code only has access to an id...

-jon