Do not decode if it was alredy decoded - i.e. if tomcat+web server. For Http we need to decode, and also for eventual protocols that send the request encoded.
Also, added an option ( defaulted to what we agreed ) to allow saving the 'original'
uri. Turning it on will bring inconsistencyes between tomcat standalone and
tomcat+web, and may be open to various security issues ( if the uri is used in any
check or processing, many encoding tricks could be played - we've been there ).
But it is the strict implementation of the 2.3 spec ( in the idea that 2.3 is
backward compat. with 2.2 ).



PR:
Obtained from:
Submitted by:
Reviewed by: