re-created functionality for mapping attributes to modules (attribute types).
now attribute groups are ordered within a module. groups that come before the dedupe check have attributes of dedupe type; groups that come after, are not dedupe attributes. you add attributes to a module by adding them to a group. attributes are ordered within a group and groups within a module. the GlobalAttributesSelect page , from which you select an attribute to add to a group, displays only attributes that have not been mapped to the current module.