Apply patch supplied by Henri Gomez.
The ajp13 connection is a persistant connection
which is created by the WebServer, Apache for exemple.

When you get many hit, Apache create many http childs
which in turn will create ajp13 connections and so
ajp13 threads.

Later Apache will kill its child and so the ajp13
connection will be dropped .

So, in early stage receiveNext(), we could receive
an IOException since the remote (WebServer) has closed
the connection.

It's not an error, and we shoudln't get an Exception here.