Important fix - this rolls back a previous change that disabled recycling.
The thread pool is used to run tasks of the same type. The tasks are initialized
only once, and the pool provides a 'free' per/thread storage ( i.e. no synchronization,
O(1) ).

The init() is not supposed to be called on each exacution - otherwise we would
lose all the recycling.

( the real fix is just one line, but I also added few flush() for better debugging
in future )