Initial support for plugable SSL implementations.
This allows Tomcat to support SSL implementation modules other than JSSE.  The previous JSSE behavior is still the default, but we are now agnostic (probably too much so) to our ServerSockets.

We probably will need better sanity checks before we go to release, but this is enough for developers.

TODO:
a) Add an sslProtocol attribute for java.protocol.handler.pkgs
b) Add a marker interface to o.a.t.util.net to enforce the isSecure attribute.