dates which are displayed in templates are now localized according to whatever locale is sent in the browser as well as having the time included in them.
in the future, we may want to make this even more customizable, but for
now, the default DateFormat.MEDIUM should do the trick.

I'm still getting ?? in the templates when the date is localized to
Korean, but i have a feeling that this is a result of either browser
issues or issues in the general localization stuff and will be resloved at
some point in the future.

-jon