updates from kevin @ bitonic...
Added in the committed-by element to the issue element.  This allows you
to specify who committed
the issue.  Added the data and created-date elements to the attachment
elements.  This allows for
comments data to be specified as attachments.  When importing from
bugzilla a comment, it is
important to specify the created date because scarab displays the created
date as the date of the
comment.

Issue number:
Obtained from:
Submitted by:
Reviewed by: