Updated UI screens:
milestone 1:
2.1.0
2.1.1
2.1.2

Milestone 3
2.1.3 - assign issue



NOTE: the stylesheets have been updated as well...  if you see a screen in the UI directory where the CSS is horked, you are looking at an old screen... we are going thru all the old stuff and making fixes... so if you need something done (e.g. prioritizes) drop me a line.

thanks