Updates so jaxpJars can include absolute paths.
Added a "jarSeparator" attribute to allow changing the separator from
the default ':' so Windows absolute paths can be used.

Implemented "additionalJars" attribute and support for an additionalJars
context property to add a list of jars to the web application classloader.
This avoids trying to use jaxpJars to perform this feature..